+++ NEWS.md
@@ -1,6 +1,7 @@
 # jlme (development version)
 
-* Bugfix passing `ftol_rel` to Julia in `parametricbootstrap()`
+* `parametricbootstrap()` gains a `optsum_overrides` argument which accepts a list of values to override in the model [OptSummary](https://juliastats.org/MixedModels.jl/stable/api/#MixedModels.OptSummary).
+* `parametricbootstrap()` now prints progress to the console.
 
 # jlme 0.2.0